refactor: LottoRanks 일급 컬렉션 추가
- TotalReturnCalculator 제거 후 책임 이전
- Test 코드 수정
