+++ README.md
@@ -29,3 +29,4 @@
 - [x] LottoNumbers 책임을 Lotto로 이전
 - [x] Lotto의 numbers 타입을 Set으로 변경
 - [x] 로또 관련 상수 LottoRules Enum Class로 관리하도록 변경 
+- [x] LottoRanks 일급 컬렉션 추가
